- qb 696: move initialOzonePrice from PoT to Register module
