+++ x/pot/genesis.go
@@ -9,7 +9,6 @@ import (
 // and the keeper's address to pubkey map
 func InitGenesis(ctx sdk.Context, keeper Keeper, data types.GenesisState) {
 	keeper.SetParams(ctx, data.Params)
-	keeper.SetInitialUOzonePrice(ctx, data.InitialUozPrice)
 }
 
 // ExportGenesis writes the current store values
@@ -17,7 +16,5 @@ func InitGenesis(ctx sdk.Context, keeper Keeper, data types.GenesisState) {
 // with InitGenesis
 func ExportGenesis(ctx sdk.Context, keeper Keeper) (data types.GenesisState) {
 	params := keeper.GetParams(ctx)
-	initialUOzonePrice := keeper.GetInitialUOzonePrice(ctx)
-
-	return types.NewGenesisState(params, initialUOzonePrice)
+	return types.NewGenesisState(params)
 }
