+++ x/pot/types/errors.go
@@ -14,14 +14,13 @@ var (
 	ErrInsufficientMatureTotal           = sdkerrors.Register(ModuleName, 7, "insufficient mature total")
 	ErrInsufficientFoundationAccBalance  = sdkerrors.Register(ModuleName, 8, "insufficient foundation account balance")
 	ErrInsufficientUnissuedPrePayBalance = sdkerrors.Register(ModuleName, 9, "insufficient unissued prepay balance")
-	ErrInitialUOzonePrice                = sdkerrors.Register(ModuleName, 10, "initial uOzone price must be positive")
-	ErrNotTheOwner                       = sdkerrors.Register(ModuleName, 11, "not the owner of the node")
-	ErrMatureEpoch                       = sdkerrors.Register(ModuleName, 12, "the value of epoch must be positive and greater than its previous one")
-	ErrEmptyFromAddr                     = sdkerrors.Register(ModuleName, 13, "missing from address")
-	ErrEmptyReporterAddr                 = sdkerrors.Register(ModuleName, 14, "missing reporter address")
-	ErrEmptyNodesVolume                  = sdkerrors.Register(ModuleName, 15, "nodes volume list empty")
-	ErrEpochNotPositive                  = sdkerrors.Register(ModuleName, 16, "report epoch is not positive")
-	ErrEmptyReportReference              = sdkerrors.Register(ModuleName, 17, "missing report reference")
-	ErrEmptyReporterOwnerAddr            = sdkerrors.Register(ModuleName, 18, "missing reporter owner address")
-	ErrNegativeVolume                    = sdkerrors.Register(ModuleName, 19, "report volume is negative")
+	ErrNotTheOwner                       = sdkerrors.Register(ModuleName, 10, "not the owner of the node")
+	ErrMatureEpoch                       = sdkerrors.Register(ModuleName, 11, "the value of epoch must be positive and greater than its previous one")
+	ErrEmptyFromAddr                     = sdkerrors.Register(ModuleName, 12, "missing from address")
+	ErrEmptyReporterAddr                 = sdkerrors.Register(ModuleName, 13, "missing reporter address")
+	ErrEmptyNodesVolume                  = sdkerrors.Register(ModuleName, 14, "nodes volume list empty")
+	ErrEpochNotPositive                  = sdkerrors.Register(ModuleName, 15, "report epoch is not positive")
+	ErrEmptyReportReference              = sdkerrors.Register(ModuleName, 16, "missing report reference")
+	ErrEmptyReporterOwnerAddr            = sdkerrors.Register(ModuleName, 17, "missing reporter owner address")
+	ErrNegativeVolume                    = sdkerrors.Register(ModuleName, 18, "report volume is negative")
 )
