+++ x/register/genesis.go
@@ -49,6 +49,7 @@ func InitGenesis(ctx sdk.Context, keeper Keeper, data types.GenesisState) {
 
 	keeper.SetInitialGenesisStakeTotal(ctx, initialStakeTotal)
 	keeper.SetRemainingOzoneLimit(ctx, initialStakeTotal)
+	keeper.SetInitialUOzonePrice(ctx, data.InitialUozPrice)
 }
 
 // ExportGenesis writes the current store values
@@ -71,6 +72,7 @@ func ExportGenesis(ctx sdk.Context, keeper Keeper) (data types.GenesisState) {
 
 	resourceNodes := keeper.GetAllResourceNodes(ctx)
 	indexingNodes := keeper.GetAllIndexingNodes(ctx)
+	initialUOzonePrice := keeper.GetInitialUOzonePrice(ctx)
 
 	return types.GenesisState{
 		Params:                 params,
@@ -78,5 +80,6 @@ func ExportGenesis(ctx sdk.Context, keeper Keeper) (data types.GenesisState) {
 		ResourceNodes:          resourceNodes,
 		LastIndexingNodeStakes: lastIndexingNodeStakes,
 		IndexingNodes:          indexingNodes,
+		InitialUozPrice:        initialUOzonePrice,
 	}
 }
