+++ x/register/keeper/keeper.go
@@ -60,6 +60,22 @@ func (k *Keeper) SetHooks(sh types.RegisterHooks) *Keeper {
 	return k
 }
 
+func (k Keeper) SetInitialUOzonePrice(ctx sdk.Context, price sdk.Dec) {
+	store := ctx.KVStore(k.storeKey)
+	b := k.cdc.MustMarshalBinaryLengthPrefixed(price)
+	store.Set(types.InitialUOzonePriceKey, b)
+}
+
+func (k Keeper) GetInitialUOzonePrice(ctx sdk.Context) (price sdk.Dec) {
+	store := ctx.KVStore(k.storeKey)
+	b := store.Get(types.InitialUOzonePriceKey)
+	if b == nil {
+		panic("Stored initial uOzone price should not have been nil")
+	}
+	k.cdc.MustUnmarshalBinaryLengthPrefixed(b, &price)
+	return
+}
+
 func (k Keeper) SetInitialGenesisStakeTotal(ctx sdk.Context, stake sdk.Int) {
 	store := ctx.KVStore(k.storeKey)
 	b := k.cdc.MustMarshalBinaryLengthPrefixed(stake)
